switching to double type

+++ examples/telemetry.c
@@ -133,7 +133,7 @@ int main(int argc, char *argv[])
   
 }
 
-int tempconverter(int tempinkelvin){
+double tempconverter(int tempinkelvin){
   double result;
   result = (((tempinkelvin/100) - 273.15) * 1.8) + 32;
   return result;
